refactor(embed): address polish review findings (#52)
Polish pass after PR #0's initial commit. Applies multi-reviewer
findings (1 high, 3 medium, 4 low) without changing the PR's scope or
the public API surface.

- REUSE-001 (high): `save` in `src/embed/fixtures.rs` now writes each
  chunk's payload via `crate::storage::f32_as_bytes` as a single
  zero-copy slice instead of looping one `f32::to_le_bytes` at a time.
  `load` mirrors the change via a single `read_exact` + `bytemuck::cast_slice`.
  Crate-wide little-endian invariant (`src/storage.rs` `compile_error!`
  guard) makes the replacement safe.
- REUSE-002 + CQ-1 (medium): workload generators extracted to
  `src/embed/workloads.rs` (new `pub mod workloads`, no feature gate).
  `mlx_smoke.rs` imports from there instead of keeping a local copy, so
  the two versions can no longer drift silently and invalidate the
  captured fixtures.
- CQ-2 (medium): unknown `mlx_smoke` subcommand now emits a warning
  before falling through to the legacy assertions. Typos like
  `measur-baseline` are no longer silent.
- CQ-5 (medium): `fixtures::cosine` returns `1.0` on identical zero
  vectors (not `0.0`) so that a bit-identical zero fixture does not
  fail NFR-001 catastrophically. Added two regression tests covering
  identical-zero and zero-vs-nonzero paths.
- EFF-001 (medium): `run_measure_baseline` now warms W1, W2, and W3
  each (one batch per shape) before the timed loop. MLX kernels are
  compiled per `(batch_size, max_seq_len)` shape and the previous
  single-W1 warmup left W2's and W3's first timed call carrying the
  compile-spike cost. Re-measured baseline numbers with the three-shape
  warmup update `docs/benchmarks/phase1_baseline.md`:
  - W1: ratio 1.130 → 1.220 (compile cost was previously hidden in
    sequential too, and the removed bias there is larger than the
    bias removed from batch)
  - W2: ratio 0.112 → 0.101 (tiny; W2 was never cold-compile-bound)
  - W3: ratio 1.926 → 1.710 (cold compile on the first timed call
    was inflating the headroom estimate — the real improvement target
    is 11% less than first reported, still a large win)
- CQ-3 (low): `FixtureDiff` field order now reads `cosine_min` first,
  matching the NFR-001 wording and the order users reason about.
- CQ-4 (low): dropped "Phase 2" phrasing from `fixtures.rs` module/
  constant docs; the spec-id reference stays evergreen.
- EFF-002 (low): `save` and `load` doc comments now explicitly require
  wrapping `File` handles in `BufWriter` / `BufReader` to avoid the
  per-chunk-boundary syscall penalty the self-describing format
  would otherwise trigger.

Test: `cargo test --workspace` 198 passed (7 fixtures + 4 workloads
+ 187 pre-existing), `cargo clippy -D warnings` clean, `cargo fmt`
clean, re-ran `mlx_smoke capture-fixture` (no fixture changes) and
`mlx_smoke measure-baseline` (new numbers reflected in
phase1_baseline.md).

Part of #52
